requirements.txt
flask
flask-bcrypt
flask-cors
flask-sqlalchemy
selenium
yfinance
